Support fixed tile size of 32 for some common kernels
Implemenations of SHFL and BALLOT depend on TILE_SIZE of the current
source (which can be 32 while Hipcontext's tileSize is 64).

+++ platforms/hip/src/HipNonbondedUtilities.cpp
@@ -200,7 +200,7 @@ static bool compareInt2LargeSIMD(int2 a, int2 b) {
 
 void HipNonbondedUtilities::initialize(const System& system) {
     if (context.getTileSize() > 32) {
-        initialize<unsigned long>(system);
+        initialize<unsigned long long>(system);
     }
     else {
         initialize<unsigned int>(system);
